✨ OGP 設定する Fixes #19
